Refactor: Predicates into separate package
Create a new package under seedu.mark.model for the predicates used in
searching.
