+++ src/main/java/seedu/mark/model/predicates/IdentifiersContainKeywordsPredicate.java
@@ -1,8 +1,10 @@
-package seedu.mark.model.bookmark;
+package seedu.mark.model.predicates;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
+
+import seedu.mark.model.bookmark.Bookmark;
 
 /**
  * Tests that a {@code Bookmark}'s {@code Name} or {@code Url} matches any
