+++ src/test/java/seedu/mark/logic/commands/CommandTestUtil.java
@@ -17,7 +17,7 @@ import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.model.Mark;
 import seedu.mark.model.Model;
 import seedu.mark.model.bookmark.Bookmark;
-import seedu.mark.model.bookmark.NameContainsKeywordsPredicate;
+import seedu.mark.model.predicates.NameContainsKeywordsPredicate;
 import seedu.mark.testutil.EditBookmarkDescriptorBuilder;
 
 /**
