+++ src/test/java/seedu/mark/model/predicates/UrlContainsKeywordsPredicateTest.java
@@ -1,4 +1,4 @@
-package seedu.mark.model.bookmark;
+package seedu.mark.model.predicates;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
